added the script file
Make sure to edit the script before running it.
Change <hostname> to the hostname of the machine that will be using the certificates.
GLHF, also i have not license and will not be held accountable in case the script ruins your day/week or server.
This script doesn't require root privileges. although running it as root shouldn't matter.
The script is designed to work with Debian based distro's although the certificates can also be used on any other machine.